Update JustEnoughAdministration.psm1
Updated ComplexObjectsEqual Method.
This method compares two JSON strings that are the result of a ConvertTo-Json from two hashtables. When these hashtables have more than one key (in case you are using more than one role defition) this could fail if order is not maintained.

This example shows the objects compared when this happens. In this case the method returns a false:

[DCTEST]: [DBG]: [Process:1916]: [Runspace3]: PS C:\Windows\system32>> ConvertTo-Json ($this.ConvertStringToHashtable($currentInstance.RoleDefinitions)) -Depth 100
{
    "GOCGROUP":  {
                     "RoleCapabilities":  "JeaCompanyDC_GOC"
                 },
    "ADMDGroup":  {
                      "RoleCapabilities":  "JEA_CompanyDC_ADMD"
                  }
}

[DCTEST]: [DBG]: [Process:1916]: [Runspace3]: PS C:\Windows\system32>> ConvertTo-Json ($roleDefinitionsHash) -Depth 100
{
    "ADMDGroup":  {
                     "RoleCapabilities":  "JEA_CompanyDC_ADMD"
                  },
    "GOCGROUP":  {
                     "RoleCapabilities":  "JeaCompanyDC_GOC"
                 }
}

Converting the hashtables into [System.Collections.Specialized.OrderedDictionary] objetcs before exporting them to JSON strings, solves the problem.